requirements.txt
Flask==3.1.1
Flask-SocketIO==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
requests==2.32.4
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0
bidict==0.23.1
simple-websocket==1.1.0
wsproto==1.2.0
h11==0.16.0
python-telegram-bot==21.7
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.35.2
dnspython==2.6.1 